config, web_util: move get_restful_params(..) to web_util
