+++ keylime/config.py
@@ -5,7 +5,6 @@ Copyright 2017 Massachusetts Institute of Technology.
 import os
 import os.path
 import configparser
-import urllib.parse
 import re
 from typing import Optional
 
@@ -16,7 +15,6 @@ except ImportError:
     from yaml import SafeLoader
 from yaml.reader import ReaderError
 
-from keylime import api_version as keylime_api_version
 from keylime import json
 
 
@@ -170,16 +168,6 @@ def ch_dir(path, logger):
     os.chdir(path)
 
 
-def list_to_dict(alist):
-    """Convert list into dictionary via grouping [k0,v0,k1,v1,...]"""
-    params = {}
-    i = 0
-    while i < len(alist):
-        params[alist[i]] = alist[i + 1] if (i + 1) < len(alist) else None
-        i = i + 2
-    return params
-
-
 def yaml_to_dict(arry, add_newlines=True, logger=None) -> Optional[dict]:
     arry = convert(arry)
     sep = "\n" if add_newlines else ""
@@ -189,28 +177,6 @@ def yaml_to_dict(arry, add_newlines=True, logger=None) -> Optional[dict]:
         if logger is not None:
             logger.warning("Could not load yaml as dict: %s", str(err))
     return None
-
-
-def get_restful_params(urlstring):
-    """Returns a dictionary of paired RESTful URI parameters"""
-    parsed_path = urllib.parse.urlsplit(urlstring.strip("/"))
-    query_params = urllib.parse.parse_qsl(parsed_path.query)
-    path_tokens = parsed_path.path.split('/')
-
-    # If first token looks like an API version, validate it and make sure it's supported
-    api_version = 0
-    if path_tokens[0] and len(path_tokens[0]) >= 0 and re.match(r"^v?[0-9]+(\.[0-9]+)?", path_tokens[0]):
-        version = keylime_api_version.normalize_version(path_tokens[0])
-
-        if keylime_api_version.is_supported_version(version):
-            api_version = version
-
-        path_tokens.pop(0)
-
-    path_params = list_to_dict(path_tokens)
-    path_params["api_version"] = api_version
-    path_params.update(query_params)
-    return path_params
 
 
 def valid_exclude_list(exclude_list):
